\name{ClusterProj-package}
\alias{ClusterProj-package}
\alias{ClusterProj}
\docType{package}
\title{
\packageTitle{ClusterProj}
}
\description{
\packageDescription{ClusterProj}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ClusterProj}
\packageIndices{ClusterProj}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ClusterProj}

Maintainer: \packageMaintainer{ClusterProj}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
